refactor build_stac_catalog.py, fix imports
