+++ scripts/build_stac_catalog.py
@@ -17,7 +17,7 @@ import stac_geoparquet
 from xopr.stac import (
     create_catalog, create_collection,
     build_collection_extent, create_items_from_flight_data,
-    discover_campaigns, discover_flight_lines
+    discover_campaigns, discover_flight_lines, build_limited_catalog
 )
 
 # STAC extension URLs
@@ -97,263 +97,6 @@ def print_catalog_structure(catalog: pystac.Catalog, indent: int = 0) -> None:
     for child in child_catalogs:
         print_catalog_structure(child, indent + 1)
 
-
-def build_limited_catalog(
-    data_root: Path,
-    output_path: Path,
-    catalog_id: str = "OPR",
-    data_product: str = "CSARP_standard",
-    extra_data_products: list[str] = ['CSARP_layer',
-                                      'CSARP_qlook'],
-    base_url: str = "https://data.cresis.ku.edu/data/rds/",
-    max_items: int = 10,
-    campaign_filter: list = None
-) -> pystac.Catalog:
-    """
-    Build STAC catalog with limits for faster processing.
-
-    Parameters
-    ----------
-    data_root : Path
-        Root directory containing campaign data
-    output_path : Path
-        Directory where catalog will be saved
-    catalog_id : str, optional
-        Catalog ID, by default "OPR"
-    data_product : str, optional
-        Primary data product to process, by default "CSARP_standard"
-    extra_data_products : List[str], optional
-        Additional data products to include, by default ['CSARP_layer', 'CSARP_qlook']
-    base_url : str, optional
-        Base URL for asset hrefs, by default "https://data.cresis.ku.edu/data/rds/"
-    max_items : int, optional
-        Maximum number of items to process, by default 10
-    campaign_filter : List[str], optional
-        Specific campaigns to process, by default None (all campaigns)
-
-    Returns
-    -------
-    pystac.Catalog
-        The built STAC catalog
-    """
-    catalog = create_catalog(catalog_id=catalog_id)
-    campaigns = discover_campaigns(data_root)
-
-    # Filter campaigns if specified
-    if campaign_filter:
-        campaigns = [c for c in campaigns if c['name'] in campaign_filter]
-
-    for campaign in campaigns:
-        campaign_path = Path(campaign['path'])
-        campaign_name = campaign['name']
-
-        print(f"Processing campaign: {campaign_name}")
-
-        try:
-            flight_lines = discover_flight_lines(
-                campaign_path, data_product,
-                extra_data_products=extra_data_products
-            )
-        except FileNotFoundError as e:
-            print(f"Warning: Skipping {campaign_name}: {e}")
-            continue
-
-        if not flight_lines:
-            print(f"Warning: No flight lines found for {campaign_name}")
-            continue
-
-        # Limit the number of flight lines processed if specified
-        if max_items is not None:
-            flight_lines = flight_lines[:max_items]
-
-        # Group items by flight_id to create flight collections
-        flight_collections = []
-        all_campaign_items = []
-
-        for flight_data in flight_lines:
-            try:
-                items = create_items_from_flight_data(
-                    flight_data, base_url, campaign_name, data_product
-                )
-
-                # Create flight collection
-                flight_id = flight_data['flight_id']
-                flight_extent = build_collection_extent(items)
-
-                # Collect scientific metadata from items for flight collection
-                dois = [
-                    item.properties.get('sci:doi') for item in items
-                    if item.properties.get('sci:doi') is not None
-                ]
-                citations = [
-                    item.properties.get('sci:citation') for item in items
-                    if item.properties.get('sci:citation') is not None
-                ]
-
-                # Check for unique values and prepare extensions
-                flight_extensions = []
-                flight_extra_fields = {}
-
-                if dois and len(np.unique(dois)) == 1:
-                    flight_extensions.append(SCI_EXT)
-                    flight_extra_fields['sci:doi'] = dois[0]
-
-                if citations and len(np.unique(citations)) == 1:
-                    if SCI_EXT not in flight_extensions:
-                        flight_extensions.append(SCI_EXT)
-                    flight_extra_fields['sci:citation'] = citations[0]
-
-                # Collect SAR metadata from items for flight collection
-                center_frequencies = [
-                    item.properties.get('sar:center_frequency')
-                    for item in items
-                    if item.properties.get('sar:center_frequency') is not None
-                ]
-                bandwidths = [
-                    item.properties.get('sar:bandwidth')
-                    for item in items
-                    if item.properties.get('sar:bandwidth') is not None
-                ]
-
-                if (center_frequencies and
-                        len(np.unique(center_frequencies)) == 1):
-                    flight_extensions.append(SAR_EXT)
-                    flight_extra_fields['sar:center_frequency'] = (
-                        center_frequencies[0]
-                    )
-
-                if bandwidths and len(np.unique(bandwidths)) == 1:
-                    if SAR_EXT not in flight_extensions:
-                        flight_extensions.append(SAR_EXT)
-                    flight_extra_fields['sar:bandwidth'] = bandwidths[0]
-
-                flight_collection = create_collection(
-                    collection_id=flight_id,
-                    description=(
-                        f"Flight {flight_id} data from {campaign['year']} "
-                        f"{campaign['aircraft']} over {campaign['location']}"
-                    ),
-                    extent=flight_extent,
-                    stac_extensions=(
-                        flight_extensions if flight_extensions else None
-                    )
-                )
-
-                # Add scientific extra fields to flight collection
-                for key, value in flight_extra_fields.items():
-                    flight_collection.extra_fields[key] = value
-
-                # Add items to flight collection
-                flight_collection.add_items(items)
-                flight_collections.append(flight_collection)
-                all_campaign_items.extend(items)
-
-                print(
-                    f"  Added flight collection {flight_id} with "
-                    f"{len(items)} items"
-                )
-
-                # Break early if we've hit the max items limit
-                if (max_items is not None and
-                        len(all_campaign_items) >= max_items):
-                    break
-
-            except Exception as e:
-                flight_id = flight_data['flight_id']
-                print(f"Warning: Failed to process flight {flight_id}: {e}")
-                continue
-
-        if flight_collections:
-            # Create campaign collection with extent covering all flights
-            campaign_extent = build_collection_extent(all_campaign_items)
-
-            # Collect scientific metadata from all campaign items
-            campaign_dois = [
-                item.properties.get('sci:doi') for item in all_campaign_items
-                if item.properties.get('sci:doi') is not None
-            ]
-            campaign_citations = [
-                item.properties.get('sci:citation')
-                for item in all_campaign_items
-                if item.properties.get('sci:citation') is not None
-            ]
-
-            # Check for unique values and prepare extensions
-            campaign_extensions = []
-            campaign_extra_fields = {}
-
-            if campaign_dois and len(np.unique(campaign_dois)) == 1:
-                campaign_extensions.append(SCI_EXT)
-                campaign_extra_fields['sci:doi'] = campaign_dois[0]
-
-            if campaign_citations and len(np.unique(campaign_citations)) == 1:
-                if SCI_EXT not in campaign_extensions:
-                    campaign_extensions.append(SCI_EXT)
-                campaign_extra_fields['sci:citation'] = campaign_citations[0]
-
-            # Collect SAR metadata from all campaign items
-            campaign_center_frequencies = [
-                item.properties.get('sar:center_frequency')
-                for item in all_campaign_items
-                if item.properties.get('sar:center_frequency') is not None
-            ]
-            campaign_bandwidths = [
-                item.properties.get('sar:bandwidth')
-                for item in all_campaign_items
-                if item.properties.get('sar:bandwidth') is not None
-            ]
-
-            if (campaign_center_frequencies and
-                    len(np.unique(campaign_center_frequencies)) == 1):
-                campaign_extensions.append(SAR_EXT)
-                campaign_extra_fields['sar:center_frequency'] = (
-                    campaign_center_frequencies[0]
-                )
-
-            if (campaign_bandwidths and
-                    len(np.unique(campaign_bandwidths)) == 1):
-                if SAR_EXT not in campaign_extensions:
-                    campaign_extensions.append(SAR_EXT)
-                campaign_extra_fields['sar:bandwidth'] = (
-                    campaign_bandwidths[0]
-                )
-
-            campaign_collection = create_collection(
-                collection_id=campaign_name,
-                description=(
-                    f"{campaign['year']} {campaign['aircraft']} flights "
-                    f"over {campaign['location']}"
-                ),
-                extent=campaign_extent,
-                stac_extensions=(
-                    campaign_extensions if campaign_extensions else None
-                )
-            )
-
-            # Add scientific extra fields to campaign collection
-            for key, value in campaign_extra_fields.items():
-                campaign_collection.extra_fields[key] = value
-
-            # Add flight collections as children of campaign collection
-            for flight_collection in flight_collections:
-                campaign_collection.add_child(flight_collection)
-
-            catalog.add_child(campaign_collection)
-
-            print(
-                f"Added campaign collection {campaign_name} with "
-                f"{len(flight_collections)} flight collections and "
-                f"{len(all_campaign_items)} total items"
-            )
-
-    output_path.mkdir(parents=True, exist_ok=True)
-    catalog.normalize_and_save(
-        root_href=str(output_path),
-        catalog_type=pystac.CatalogType.SELF_CONTAINED
-    )
-
-    print(f"Catalog saved to {output_path}")
-    return catalog
 
 
 def export_to_geoparquet(catalog: pystac.Catalog, output_file: Path) -> None:
