+++ src/xopr/stac/__init__.py
@@ -8,7 +8,8 @@ across radar campaigns and data products.
 
 from .catalog import (
     create_catalog, create_collection, create_item,
-    build_collection_extent, create_items_from_flight_data
+    build_collection_extent, create_items_from_flight_data,
+    build_limited_catalog
 )
 from .metadata import extract_item_metadata, discover_campaigns, discover_flight_lines
 
@@ -18,6 +19,7 @@ __all__ = [
     "create_item",
     "build_collection_extent",
     "create_items_from_flight_data",
+    "build_limited_catalog",
     "extract_item_metadata",
     "discover_campaigns",
     "discover_flight_lines"
